+++ packages/apps/esm-login-app/webpack.config.js
@@ -1,1 +1,2 @@
-module.exports = require("openmrs/default-webpack-config");
+// module.exports = require("openmrs/default-webpack-config");
+module.exports = require("./webpack");
