+++ packages/shell/esm-app-shell/src/index.ejs
@@ -14,7 +14,7 @@
     <script type="systemjs-importmap">
       {
         "imports": {
-          "@openmrs/esm-patient-chart-app": "//localhost:8081/openmrs-esm-patient-chart-app.js"
+          "@openmrs/esm-devtools-app": "//localhost:8081/openmrs-esm-devtools-app.js"
         }
       }
     </script>
